requirements.txt
azure-ai-projects 
azure-ai-inference 
azure-identity
python-dotenv
gradio